feat(library): refactor library name
