+++ app/build.gradle.kts
@@ -45,5 +45,5 @@ dependencies {
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
-    implementation(project(":mark_biometric"))
+    implementation(project(":mark_authenticator"))
 }
