+++ mark_authenticator/proguard-rules.pro
