+++ mark_authenticator/src/androidTest/java/com/fadlurahmanfdev/mark_authenticator/ExampleInstrumentedTest.kt
@@ -1,0 +1,25 @@
+package com.fadlurahmanfdev.mark_authenticator
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert
+import org.junit.Test
+import org.junit.runner.RunWith
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
+        Assert.assertEquals(
+            "com.fadlurahmanfdev.kotlin_feature_identity.test",
+            appContext.packageName
+        )
+    }
+}
