+++ mark_authenticator/src/main/AndroidManifest.xml
