+++ mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/core/enums/MarkAuthenticationType.kt
@@ -1,0 +1,7 @@
+package com.fadlurahmanfdev.mark_authenticator.core.enums
+
+enum class MarkAuthenticationType {
+    BIOMETRIC_WEAK,
+    BIOMETRIC_STRONG,
+    DEVICE_CREDENTIAL,
+}
