+++ mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/core/enums/MarkAuthenticatorMethod.kt
@@ -1,0 +1,13 @@
+package com.fadlurahmanfdev.mark_authenticator.core.enums
+
+enum class MarkAuthenticatorMethod {
+    /**
+     * Represents biometric authentication using weak biometric
+     */
+    BIOMETRIC,
+
+    /**
+     * Represents biometric authentication using device credential (PIN, PASSWORD, or PATTERN).
+     */
+    DEVICE_CREDENTIAL
+}
