+++ mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/core/exception/FeatureIdentityException.kt
@@ -1,0 +1,6 @@
+package com.fadlurahmanfdev.mark_authenticator.core.exception
+
+data class FeatureIdentityException(
+    val code: String,
+    override val message: String? = null,
+) : Throwable(message = message)
