+++ mark_authenticator/src/test/java/com/fadlurahmanfdev/mark_authenticator/ExampleUnitTest.kt
@@ -1,0 +1,16 @@
+package com.fadlurahmanfdev.mark_authenticator
+
+import org.junit.Assert
+import org.junit.Test
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        Assert.assertEquals(4, 2 + 2)
+    }
+}
