--- mark_biometric/src/androidTest/java/com/fadlurahmanfdev/mark_biometric/ExampleInstrumentedTest.kt
@@ -1,25 +1,0 @@
-package com.fadlurahmanfdev.mark_biometric
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import org.junit.Assert
-import org.junit.Test
-import org.junit.runner.RunWith
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        Assert.assertEquals(
-            "com.fadlurahmanfdev.kotlin_feature_identity.test",
-            appContext.packageName
-        )
-    }
-}
