--- mark_biometric/src/main/java/com/fadlurahmanfdev/mark_biometric/core/enums/CheckAuthenticationStatusType.kt
@@ -1,7 +1,0 @@
-package com.fadlurahmanfdev.mark_biometric.core.enums
-
-enum class CheckAuthenticationStatusType {
-    BIOMETRIC_WEAK,
-    BIOMETRIC_STRONG,
-    DEVICE_CREDENTIAL,
-}
