--- mark_biometric/src/main/java/com/fadlurahmanfdev/mark_biometric/core/exception/FeatureIdentityException.kt
@@ -1,6 +1,0 @@
-package com.fadlurahmanfdev.mark_biometric.core.exception
-
-data class FeatureIdentityException(
-    val code: String,
-    override val message: String? = null,
-) : Throwable(message = message)
